chore: refactor topic page
