+++ frontend/src/pages/topics/modals/record-view-modal.tsx
